+++ js/downloadbutton/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @gradio/downloadbutton
 
+## 0.1.28-beta.4
+
+### Dependency updates
+
+- @gradio/button@0.3.0-beta.4
+
+## 0.1.28-beta.3
+
+### Dependency updates
+
+- @gradio/client@1.6.0-beta.3
+- @gradio/button@0.3.0-beta.3
+
+## 0.1.28-beta.2
+
+### Dependency updates
+
+- @gradio/client@1.6.0-beta.2
+- @gradio/utils@0.7.0-beta.2
+- @gradio/button@0.3.0-beta.2
+
+## 0.1.28-beta.1
+
+### Dependency updates
+
+- @gradio/utils@0.7.0-beta.1
+- @gradio/client@1.6.0-beta.1
+- @gradio/button@0.3.0-beta.1
+
 ## 0.1.28
 
 ### Fixes